+++ tests/torchtune/datasets/test_samsum_dataset.py
@@ -20,7 +20,7 @@ class TestSamsumDataset:
     def tokenizer(self):
         return DummyTokenizer()
 
-    @patch("torchtune.datasets._finetune.load_dataset")
+    @patch("torchtune.datasets._sft.load_dataset")
     def test_label_no_masking(self, load_dataset, tokenizer):
         """
         Test whether the input and the labels are correctly created when the input is not masked.
@@ -76,7 +76,7 @@ class TestSamsumDataset:
         ]
         assert labels == input
 
-    @patch("torchtune.datasets._finetune.load_dataset")
+    @patch("torchtune.datasets._sft.load_dataset")
     def test_label_masking(self, load_dataset, tokenizer):
         """
         Test whether the input and the labels are correctly created when the input is masked.
