+++ torchtune/datasets/_grammar.py
@@ -12,8 +12,8 @@ from torchtune.data._prompt_templates import (
     GrammarErrorCorrectionTemplate,
     PromptTemplate,
 )
-from torchtune.datasets._finetune import SFTDataset
 from torchtune.datasets._packed import PackedDataset
+from torchtune.datasets._sft import SFTDataset
 from torchtune.modules.transforms import Transform
