+++ torchtune/datasets/_slimorca.py
@@ -7,9 +7,9 @@
 from typing import Dict, Optional, Union
 
 from torchtune.data import PromptTemplate, ShareGPTToMessages
-
-from torchtune.datasets._finetune import SFTDataset
 from torchtune.datasets._packed import PackedDataset
+
+from torchtune.datasets._sft import SFTDataset
 from torchtune.modules.transforms import Transform
